Move 'Color' into its own module
