+++ src/gameboy/gfx/mod.rs
@@ -1,7 +1,8 @@
 
+mod color;
 mod frame;
 mod gpu;
 
+pub use self::color::Color;
 pub use self::gpu::Gpu;
-pub use self::frame::Color;
 pub use self::frame::Frame;
